Move partner auto-reject onto defineJob.
