+++ apps/web/lib/actions/partners/create-program-application.ts
@@ -11,6 +11,7 @@ import { getApplicationEventCookieName } from "@/lib/application-events/utils";
 import { getSession } from "@/lib/auth";
 import { qstash } from "@/lib/cron";
 import { autoApprovePartnerJob } from "@/lib/jobs/handlers/auto-approve-partner-job";
+import { autoRejectPartnerJob } from "@/lib/jobs/handlers/auto-reject-partner-job";
 import { getNetworkProfileChecklistProgress } from "@/lib/network/get-network-profile-checklist-progress";
 import { evaluateApplicationRequirements } from "@/lib/partners/evaluate-application-requirements";
 import {
@@ -279,25 +280,16 @@ async function createApplicationAndEnrollment({
       );
     }
 
-    const qstashResponse = await qstash.publishJSON({
-      url: `${APP_DOMAIN_WITH_NGROK}/api/cron/partners/auto-reject`,
-      delay: 30 * 60, // 30 minutes
-      body: {
+    await autoRejectPartnerJob.dispatch(
+      {
         programId: program.id,
         partnerId: partner.id,
       },
-    });
-
-    if (qstashResponse.messageId) {
-      console.log(
-        `The partner did not meet the eligibility requirements for this program. Auto-reject job enqueued successfully.`,
-        {
-          ...qstashResponse,
-          programId: program.id,
-          partnerId: partner.id,
-        },
-      );
-    }
+      {
+        delay: 30 * 60, // 30 minutes
+        label: partner.id,
+      },
+    );
   }
 
   const applicationId = createId({ prefix: "pga_" });
