+++ apps/web/lib/jobs/registry.ts
@@ -42,6 +42,11 @@ const jobLoaders = {
     import("./handlers/auto-approve-partner-job").then(
       (m) => m.autoApprovePartnerJob,
     ),
+
+  "auto-reject-partner-job": () =>
+    import("./handlers/auto-reject-partner-job").then(
+      (m) => m.autoRejectPartnerJob,
+    ),
 } as const satisfies Record<string, () => Promise<JobDefinition>>;
 
 const jobCache = new Map<string, JobDefinition>();
